AUD-143: coyote time, buffer de salto, anticipacion y tres tipos de camara
Se pidio verificar si estas cuatro cosas existen y llegan al jugador.
Comprobado contra el codigo, una por una:

  - COYOTE TIME: existia y funcionaba, pero se contaba en FOTOGRAMAS. Seis
    fotogramas son 100 ms a 60 fps, 200 ms a 30 y 42 ms a 144: el margen de
    perdon para saltar tarde cambiaba con la maquina, y en un portatil que
    baja a 30 fps el juego se volvia notablemente mas blando. Ahora se
    acumula con dt y dura lo mismo en cualquier equipo.
  - BUFFER DE SALTO: existia, integrado y correcto. 8 fotogramas de margen
    para adelantarse al aterrizaje. Sin cambios; ahora con pruebas.
  - ANTICIPACION: existia en horizontal y SIN SUAVIZAR. Al girar, el objetivo
    de la camara saltaba de +60 a -60 px en un fotograma, que se ve como un
    tiron. Ahora se suaviza, y ademas mira hacia abajo al caer -que es lo que
    evita el salto de fe-. Hacia arriba no mira: ya sabes de donde vienes, y
    marea al encadenar saltos.
  - TIPOS DE CAMARA: habia UNO. Ahora hay tres, con la propiedad `camara`:
    `seguir` (el de siempre), `zona_muerta` (no se mueve mientras estes en el
    centro; Celeste, Hollow Knight) y `sala` (salta de pantalla en pantalla
    sin suavizar; Zelda, Metroid, Castlevania). El corte de `sala` es
    instantaneo a proposito: suavizarlo lo convierte en un barrido y se pierde
    lo que aporta.

Y dos defectos que apareceron al mirar:

  1. UN CameraLock CONGELABA EL NIVEL ENTERO. `set_camera_locks` hacia
     `any(line.lock_x for line in locks)`: el `rect` de cada zona se guardaba
     y NO SE LEIA NUNCA, asi que una sola zona en cualquier esquina del mapa
     dejaba la camara clavada desde el primer fotograma.

     No es una sospecha: `boss_rey_scene.py` llevaba escrito un parche para
     rodearlo, tocando `_is_locked_x` desde fuera, con un comentario que
     explicaba el defecto. Cuando un escenario tiene que parchear el motor,
     el defecto es del motor. Arreglado en `Camera` y parche retirado.

  2. LA ORDEN `temblor` DE LAS CUTSCENES NO SACUDIA NADA. Mi accion de
     AUD-136 llamaba a `camera.shake()`, y el metodo se llama `apply_shake`:
     el getattr devolvia None y se salia en silencio. Mi prueba pasaba porque
     use una camara de mentira con un metodo `shake` que me invente yo. Es el
     mismo fallo que este proyecto ya cometio con el doble de `InputManager`:
     un doble que no se parece al objeto real no prueba nada. Hay una prueba
     nueva que mira la clase de verdad.

28 pruebas nuevas, y tres de ellas nacieron rotas por mi culpa:

  - la del coyote a 30 vs 144 fps pasaba VACIA, porque el jugador nace con el
    contador agotado y el bucle no daba ni una vuelta: media cero contra cero;
  - la del buffer usaba una lista de colision vacia, asi que el jugador nunca
    estaba apoyado y el buffer no podia dispararse: media la ausencia de
    suelo, no el buffer;
  - la de centrado esperaba la esquina del rect en vez de su centro.

Las tres corregidas y comprobadas contra el comportamiento real antes de
darlas por buenas.

+++ src/framework/entities/player_state.py
@@ -20,7 +20,9 @@ if TYPE_CHECKING:
 @dataclass
 class PlayerStateData:
     # Physics
-    coyote_counter: int = 0
+    #: AUD-143: en «fotogramas a 60 fps», pero acumulado con `dt`. Es
+    #: float porque el tiempo real no cae en números enteros.
+    coyote_counter: float = 0.0
     jump_cut_applied: bool = False
 
     # Attack state
